Deleted private string name and functions setName and getName in Product due to compilation errors.

+++ OOP-FriendFunctions/Rectangle.cpp
@@ -12,7 +12,7 @@ Rectangle::Rectangle(Point a, Point b, Point c, Point d) {
 	points[3] = d;
 }
 
-int length(Point p1, Point p2) {
+double length(Point p1, Point p2) {
 	return sqrt(pow((p2.getX() - p1.getX()), 2) + pow((p2.getY() - p1.getY()), 2));
 }
